+++ modules/column-articles/exemples.php
@@ -1,0 +1,25 @@
+<?php
+return [
+  'Standard' => [
+    'items' => [
+      [
+        'title' => "OPTIMA XV - 2016 : bilan d’un colloque réussi ",
+        'href' => '#',
+				'text' => "OPTIMA a tenu son 15 ème colloque sur les avancées scientifiques dans le domaine des plantes autour du bassin méditerranéen...",
+        'date' => 'il y a 2 heures'
+      ],
+      [
+        'title' => "Un(e) chargé(e) de missions « plan de gestion » pour Bretagne Vivante... ",
+        'href' => '#',
+        'text' => 'L’association Bretagne Vivante - SEPNB recherche un(e) chargé(e) de missions en CDD de 14 mois.',
+        'date' => "aujourd'hui"
+      ],
+      [
+        'title' => "À la rencontre d’un agriculteur à la retraite passionné de botanique ! ",
+        'href' => '#',
+				'text' => "Zoom sur un ancien agriculteur conventionnel aujourd’hui passionné de botanique. Comment la passion pour...",
+        'date' => 'lundi 29 août'
+      ]
+    ]
+  ]
+];